Split provider implementation into module
This commit split all the original provider implementation into
different module. This can keep the root module file tidy and readble.
It also help maintain original provider, remove old provider, extend
provider functinality and implement new providder easier.
