+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub mod data;
 pub mod helper;
-pub mod model;
 pub mod provider;
